requirements.txt
# ---------------- Core Framework ----------------
fastapi==0.110.0
uvicorn[standard]==0.27.1

# ---------------- PDF + Data ----------------
pdfplumber==0.10.3
pandas==2.2.0

# ---------------- LLM (Google Gemini) ----------------
langchain==0.1.16
langchain-google-genai==1.0.3

# ---------------- Environment ----------------
python-dotenv==1.0.1
